- 03-05-2005 changed: Get/SeekPercentage now take floats.   - 03-05-2005 added: Initial seek bar support - not currently implemented at the skin level.   - 03-05-2005 changed: Increased PAPlayer output buffer size + added debug output if it drops down.
git-svn-id: https://xbmc.svn.sourceforge.net/svnroot/xbmc/trunk/XBMC@3369 568bbfeb-2a22-0410-94d2-cc84cf5bfa90

+++ xbmc/GUIDialogProgress.cpp
@@ -127,7 +127,7 @@ void CGUIDialogProgress::SetPercentage(int iPercentage)
   if (iPercentage < 0) iPercentage = 0;
   if (iPercentage > 100) iPercentage = 100;
   CGUIProgressControl* pControl = (CGUIProgressControl*)GetControl(CONTROL_PROGRESS_BAR);
-  if (pControl) pControl->SetPercentage(iPercentage);
+  if (pControl) pControl->SetPercentage((float)iPercentage);
 }
 void CGUIDialogProgress::ShowProgressBar(bool bOnOff)
 {
